Minor fixes for compiling on MacOS

+++ jlm/llvm/opt/RvsdgTreePrinter.cpp
@@ -145,7 +145,7 @@ RvsdgTreePrinter::AnnotateNumNodes(
           numSubregionNodes += annotateRegion(*subregion);
         }
 
-        annotationMap.AddAnnotation(structuralNode, { label, numSubregionNodes });
+        annotationMap.AddAnnotation(structuralNode, { label, uint64_t(numSubregionNodes) });
       }
 
       if (match(node))
@@ -154,7 +154,7 @@ RvsdgTreePrinter::AnnotateNumNodes(
       }
     }
 
-    annotationMap.AddAnnotation(&region, { label, numNodes });
+    annotationMap.AddAnnotation(&region, { label, uint64_t(numNodes) });
 
     return numNodes;
   };
